+++ GRDB/Record/TableMapping.swift
@@ -7,6 +7,12 @@
         #else
             import SQLiteiPhoneOS
         #endif
+    #elseif os(watchOS)
+        #if (arch(i386) || arch(x86_64))
+            import SQLiteWatchSimulator
+        #else
+            import SQLiteWatchOS
+        #endif
     #endif
 #endif
 
@@ -121,7 +127,8 @@ extension TableMapping {
     
     // MARK: - Deleting by Single-Column Primary Key
     
-    /// Delete records identified by their primary keys.
+    /// Delete records identified by their primary keys; returns the number of
+    /// deleted rows.
     ///
     ///     try Person.deleteAll(db, keys: [1, 2, 3])
     ///
@@ -137,7 +144,8 @@ extension TableMapping {
         return db.changesCount
     }
     
-    /// Delete a record, identified by its primary key.
+    /// Delete a record, identified by its primary key; returns whether a
+    /// database row was deleted.
     ///
     ///     try Person.deleteOne(db, key: 123)
     ///
@@ -193,9 +201,10 @@ extension RowConvertible where Self: TableMapping {
 
     // MARK: - Fetching by Key
     
-    /// Returns a sequence of records, given an array of key dictionaries.
+    /// Returns a sequence of records identified by the provided unique keys
+    /// (primary key or any key with a unique index on it).
     ///
-    ///     let persons = Person.fetch(db, keys: [["name": "Arthur"], ["name": "Barbara"]]) // DatabaseSequence<Person>
+    ///     let persons = Person.fetch(db, keys: [["email": "a@example.com"], ["email": "b@example.com"]]) // DatabaseSequence<Person>
     ///
     /// The order of records in the returned sequence is undefined.
     ///
@@ -211,9 +220,10 @@ extension RowConvertible where Self: TableMapping {
         return fetch(statement)
     }
     
-    /// Returns an array of records, given an array of key dictionaries.
+    /// Returns an array of records identified by the provided unique keys
+    /// (primary key or any key with a unique index on it).
     ///
-    ///     let persons = Person.fetchAll(db, keys: [["name": "Arthur"], ["name": "Barbara"]]) // [Person]
+    ///     let persons = Person.fetchAll(db, keys: [["email": "a@example.com"], ["email": "b@example.com"]]) // [Person]
     ///
     /// The order of records in the returned array is undefined.
     ///
@@ -229,7 +239,8 @@ extension RowConvertible where Self: TableMapping {
         return fetchAll(statement)
     }
     
-    /// Returns a single record given a key dictionary.
+    /// Returns a single record identified by a unique key (the primary key or
+    /// any key with a unique index on it).
     ///
     ///     let person = Person.fetchOne(db, key: ["name": Arthur"]) // Person?
     ///
@@ -285,9 +296,10 @@ extension TableMapping {
 
     // MARK: - Deleting by Key
     
-    /// Delete records, given an array of key dictionaries.
+    /// Delete records identified by the provided unique keys (primary key or
+    /// any key with a unique index on it); returns the number of deleted rows.
     ///
-    ///     try Person.deleteAll(db, keys: [["name": "Arthur"], ["name": "Barbara"]])
+    ///     try Person.deleteAll(db, keys: [["email": "a@example.com"], ["email": "b@example.com"]])
     ///
     /// - parameters:
     ///     - db: A database connection.
@@ -301,7 +313,8 @@ extension TableMapping {
         return db.changesCount
     }
     
-    /// Delete a record, identified by a key dictionary.
+    /// Delete a record, identified by a unique key (the primary key or any key
+    /// with a unique index on it); returns whether a database row was deleted.
     ///
     ///     Person.deleteOne(db, key: ["name": Arthur"])
     ///